Don't try to tag the GlobalVariable itself
There isn't an "original pointer" to tag; this is a constant substituted in
by the linker in lots of places. Those places don't need tagging, because
they are e.g. assignment of a constant to a register; if we then write it to
a variable, it will be picked up and tagged.