For load/store, identify what it is really writing to.
There are two different kinds of shouldTag*. For a pointer, in general, we
need to identify whether it is sensitive, that is, whether it has any
components that are sensitive. However, for a load/store operation, we need
to check what we are actually writing to. In particular, writing a word to
an object is likely to be the vptr, and we need to unpack the structure to
figure this out. This is different for initializers.